backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
